Add files via upload
Added a shopping cart counter so it displays how many items are in cart (upon refresh) on the navigation bar. Also added a "Product already in cart" if you try to add an item again. Changed the CSS so it displays as purple/blue and down a little bit more.